[REFACTOR]:[工程结构优化]
变更说明：
1.pom.xml文件中依赖包及插件版本等属性统一在properties中管理
2.添加lombok依赖，简化代码
3.添加developer信息
