+++ src/main/java/com/alibaba/dubbo/spring/boot/configuration/DubboSpringBootStarterConstants.java
@@ -1,4 +1,6 @@
-package com.alibaba.dubbo.spring.boot.domain;
+package com.alibaba.dubbo.spring.boot.configuration;
+
+import lombok.experimental.UtilityClass;
 
 /**
  * Constants defined.
@@ -7,6 +9,7 @@ package com.alibaba.dubbo.spring.boot.domain;
  * @version 1.0.0
  * @since 1.0.0
  */
+@UtilityClass
 public class DubboSpringBootStarterConstants {
 
   public static final String GROUP = "group";
