+++ src/main/java/com/alibaba/dubbo/spring/boot/listener/ConsumerSubscribeListener.java
@@ -9,7 +9,7 @@ import com.alibaba.dubbo.rpc.Invoker;
 import com.alibaba.dubbo.rpc.RpcException;
 import com.alibaba.dubbo.rpc.listener.InvokerListenerAdapter;
 import com.alibaba.dubbo.spring.boot.domain.ClassIdBean;
-import com.alibaba.dubbo.spring.boot.domain.DubboSpringBootStarterConstants;
+import com.alibaba.dubbo.spring.boot.configuration.DubboSpringBootStarterConstants;
 
 /**
  * Dubbo client invoker listener
